#14 - feat : TitleString을 core:ui에 정의
그러므로 MainScreen과 SignInScreen에서 정의한것은 지움
