+++ feature/signin/src/main/java/doingwell/feature/signin/SignInRootScreen.kt
@@ -16,16 +16,12 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.hegunhee.youredoingwell.ui.theme.MainGreen
 import com.hegunhee.youredoingwell.ui.theme.Typography
+import doingwell.core.ui.text.TitleText
 
 @Composable
 fun SignInRootScreen(
@@ -46,9 +42,8 @@ fun SignInScreen(
         modifier = modifier.fillMaxWidth()
     ) {
         Spacer(modifier = modifier.padding(top = 100.dp))
-        Text(
-            titleString(),
-            style = Typography.titleLarge,
+        TitleText(
+            textStyle = Typography.titleLarge,
         )
 
         Spacer(modifier = modifier.padding(top = 15.dp))
@@ -91,20 +86,6 @@ fun SignInScreen(
             Spacer(modifier = modifier.padding(horizontal = 10.dp))
             Text(stringResource(R.string.sign_up))
         }
-    }
-}
-
-private fun titleString(): AnnotatedString {
-    return buildAnnotatedString {
-        withStyle(style = SpanStyle(color = MainGreen)) {
-            append("당")
-        }
-        append("신은")
-        append("\n")
-        withStyle(style = SpanStyle(color = MainGreen)) {
-            append("잘")
-        }
-        append("하고 있다")
     }
 }
